+++ src/Spinorama/index.ts
@@ -1,0 +1,2 @@
+export { Spinorama } from './Spinorama'
+export * from './Spinorama.types'