+++ src/index.ts
@@ -1,1 +1,2 @@
+export * from './Spinorama'
 export * from './Slideshow'