+++ src/index.d.ts
@@ -1,2 +1,2 @@
 export * from './Spinorama'
-export * from './Slideshow'
+export * from './SpinoramaSlideshow'