+++ src/Slideshow/index.ts
@@ -1,0 +1,2 @@
+export { SpinoramaSlideshow } from './Slideshow'
+export * from './Slideshow.types'