+++ src/SpinoramaSlideshow/index.ts
@@ -1,2 +1,0 @@
-export { SpinoramaSlideshow } from './SpinoramaSlideshow'
-export * from './SpinoramaSlideshow.types'