+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Change the name of `SpinoramaSlideshow` to `Slideshow`
+
 ## 0.1.11
 - Export `types` in the output