+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this package will be documented in this file.
 
 - Change `package.json`
+- Change the name of `Slideshow` to `SpinoramaSlideshow`
 
 ## 0.1.12
 - Change the name of `SpinoramaSlideshow` to `Slideshow`